hub/requirements.txt
pydantic==2.9.1
fastapi==0.115.2
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
pgvector==0.2.5
celery==5.4.0
redis==5.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
beautifulsoup4==4.12.3
tenacity==9.0.0
python-slugify==8.0.4
python-dotenv==1.0.1
openai==1.51.0
